requirements.txt
websockets
opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-gcp-monitoring
